updates
- new routes
- edited Pages.php
- some fixes for this update

+++ application/classes/Controller/Pages.php
@@ -14,7 +14,7 @@ class Controller_Pages extends Controller_Base_preDispatch
         {
             if (!$uri)
             {
-                $this->redirect('/page/' . $page->id . '/' . $page->uri);
+                $this->redirect('/' . $page->id . '/' . $page->uri);
             }
 
             if ($page->id_parent)
@@ -29,91 +29,67 @@ class Controller_Pages extends Controller_Base_preDispatch
             $this->template->content = View::factory('templates/page', $this->view);
 
         } else {
-            # TODO ошибка: статья не найдена
+
+            self::error_page('Запрашиваемая страница не найдена');
+            return FALSE;
         }
     }
 
-    public function action_news_add()
+    public function action_add_new()
     {
-        if (!$this->user->isAdmin()) {
-            # TODO ошибка: недостаточно прав
-            $this->redirect('/');
+        if (!$this->user->isTeacher())
+        {
+            self::error_page('Недостаточно прав для совершения данного действия');
+            return FALSE;
         }
 
         $page = new Model_Page();
 
-        $page->type = Model_Page::TYPE_SITE_NEWS;
+        $type = $this->request->param('type');
 
-        if (Security::check(Arr::get($_POST, 'csrf')))
+        if (isset($type))
         {
-            $url = self::form_to_model();
+            switch ($this->request->param('type')) {
+                case 'news' :
+                    $page->type = Model_Page::TYPE_SITE_NEWS;
+                    break;
+                case 'page' :
+                default :
+                    $page->type = Model_Page::TYPE_USER_PAGE;
+            }
+        } else {
 
-            if ($url)
+            $page->id_parent = $this->request->param('id');
+
+            if ($page->id_parent)
             {
-                $this->redirect($url);
+                $page->parent = Model_Page::get($page->id_parent);
+
+                switch ($page->parent->type)
+                {
+                    case Model_Page::TYPE_USER_PAGE :
+                                $page_type = Model_Page::TYPE_USER_PAGE; break;
+                    default :   $page_type = Model_Page::TYPE_SITE_PAGE;
+                }
+                $page->type = $page_type;
             }
         }
 
-        $this->view['page']      = $page;
-        $this->template->content = View::factory('templates/page_form', $this->view);
-
-    }
-
-    public function action_page_add()
-    {
-        if (!$this->user->isTeacher) {
-            # TODO ошибка: недостаточно прав
-            $this->redirect('/');
-        }
-
-        $page = new Model_Page();
-
-        $page->type = Model_Page::TYPE_USER_PAGE;
-
-        if (Security::check(Arr::get($_POST, 'csrf')))
+        if (!$this->user->isAdmin() && $page->type != Model_Page::TYPE_USER_PAGE)
         {
-            $url = self::form_to_model();
-
-            if ($url)
-            {
-                $this->redirect($url);
-            }
-        }
-
-        $this->view['page']      = $page;
-        $this->template->content = View::factory('templates/page_form', $this->view);
-    }
-
-    public function action_subpage_add()
-    {
-        if (!$this->user->isTeacher) {
-            # TODO ошибка: недостаточно прав
-            $this->redirect('/');
-        }
-
-        $page = new Model_Page();
-        $page->id_parent = $this->request->param('id');
-
-        if ($page->id_parent)
-        {
-            $page->parent = Model_Page::get($page->id_parent);
-
-            switch ($page->parent->type)
-            {
-                case Model_Page::TYPE_USER_PAGE :
-                            $page_type = Model_Page::TYPE_USER_PAGE; break;
-                default :   $page_type = Model_Page::TYPE_SITE_PAGE;
-            }
-            $page->type = $page_type;
+            self::error_page('Недостаточно прав для совершения данного действия');
+            return FALSE;
         }
 
         if (Security::check(Arr::get($_POST, 'csrf')))
         {
-            $url = self::form_to_model();
+            $page = self::get_form();
 
-            if ($url)
+            if ($page->title && $page->type)
             {
-                $this->redirect($url);
+                $this->redirect(self::save_page($page));
+            } else {
+                $page->uri = 'no-title';
             }
         }
 
@@ -130,19 +106,23 @@ class Controller_Pages extends Controller_Base_preDispatch
         {
             if (Security::check(Arr::get($_POST, 'csrf')))
             {
-                $url = self::form_to_model();
-                
-                if ($url)
+                $page = self::get_form();
+
+                if ($page->title && $page->type)
                 {
-                    $this->redirect($url);
-                }    
+                    $this->redirect(self::save_page($page));
+                } else {
+                    $page->uri = 'no-title';
+                }
             }
 
             $this->view['page'] = $page;
             $this->template->content = View::factory('templates/page_form', $this->view);
         } else {
-            # TODO ошибка: недостаточно прав
-            $this->redirect('/');
+
+            self::error_page('Недостаточно прав для совершения данного действия');
+            return FALSE;
+
         }
     }
 
@@ -151,65 +131,59 @@ class Controller_Pages extends Controller_Base_preDispatch
         $id = $this->request->param('id');
         $page = Model_Page::get($id);
 
-        if ($this->user->isAdmin || $this->user->id == $page->author->id)
+        if ($this->user->isAdmin || ($this->user->id == $page->author->id && $this->user->isTeacher))
         {
             $page->delete();
 
             # правила редиректа
             if ($page->id_parent != 0) {
-                $url = '/page/' . $page->id_parent;
+                $url = '/' . $page->id_parent;
             } elseif ($page->type != Model_Page::TYPE_SITE_NEWS) {
                 $url = '/user/' . $page->author->id;
             } else {
                 $url = '/';
             }
         } else {
-            # TODO ошибка: недостаточно прав
-            $url = '/';
+
+            self::error_page('Недостаточно прав для совершения данного действия');
+            return FALSE;
+
         }
 
         $this->redirect($url);
     }
 
-    public function form_to_model()
+    public function get_form()
     {
-        $id     = (int)Arr::get($_POST, 'id');
-        $type   = (int)Arr::get($_POST, 'type');
+        $page = new Model_Page();
 
-        if ($type) {
-            $page = new Model_Page();
-            $page->type          = $type;
-            $page->author        = $this->user;
-            $page->id_parent     = (int)Arr::get($_POST, 'id_parent', 0);
-            $page->title         = Arr::get($_POST, 'title');
-            $page->content       = Arr::get($_POST, 'content');
-            $page->is_menu_item  = Arr::get($_POST, 'is_menu_item', 0);
+        $page->id            = (int)Arr::get($_POST, 'id');
+        $page->type          = (int)Arr::get($_POST, 'type');
+        $page->author        = $this->user;
+        $page->id_parent     = (int)Arr::get($_POST, 'id_parent', 0);
+        $page->title         = Arr::get($_POST, 'title', '');
+        $page->content       = Arr::get($_POST, 'content', '');
+        $page->is_menu_item  = Arr::get($_POST, 'is_menu_item', 0);
 
-            if ($page->title) {
-                return self::save_page($page, $id);
-            } else {
-                return FALSE;
-                # TODO ошибка: отсутствует заголовок
-            }
-        } else {
-            return FALSE;
-            # TODO ошибка: отсутствует тип страницы
-        }
-        
+        return $page;
     }
     
-    public function save_page($page, $id)
+    public function save_page($page)
     {
-        if ($id) {
-            $page->id = $id;
+        if ($page->id) {
             $page->update();
-            $url = '/page/' . $id;
         } else {
             $page = $page->insert();
-            $url = '/page/' . $page->id;
         }
         
-        return $url;
+        return '/' . $page->id;
+    }
+
+    public function error_page($error_text)
+    {
+        $this->view['error_text'] = $error_text;
+
+        $this->template->content = View::factory('templates/error', $this->view);
     }
